+++ cli/CHANGELOG.md
@@ -15,7 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - All ledgers can be seen
   - All payments can be seen
   - All trades can be seen
-  - Ledger Transactions can be seen
+  - Ledger effects can be seen
+  - Ledger transactions can be seen
   - Ledger payments can be seen
   - Trade aggregations can be seen
   - Transaction details can be seen